Move some plots from rbfe.py to plots.py
